requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
sockets==1.0.0
Werkzeug==1.0.1
gunicorn==20.0.4
argon2-cffi==20.1.0